requirements.txt
requests
paho-mqtt<2.0.0
jsonpath-ng
